harness/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoform-eval"
version = "0.1.0"
description = "NL to Lean autoformalization evaluation harness"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
gemini = ["google-genai>=0.4.0"]

[project.scripts]
autoform-eval = "autoform_eval.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
